beore long

+++ Sep-3-main/Server/src/main/java/com/ServerSep3/Server/Model/PostModel.java
@@ -7,7 +7,7 @@ import jakarta.persistence.*;
 public class PostModel {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
     @ManyToOne
     @JoinColumn(name = "user_id")
     private UserModel user;
@@ -27,7 +27,7 @@ public class PostModel {
     public PostModel() {
     }
 
-    public PostModel(int id, UserModel user, CategoryModel category, String title, String description, String imageUrl, LocationModel location) {
+    public PostModel(Long id, UserModel user, CategoryModel category, String title, String description, String imageUrl, LocationModel location) {
         this.id = id;
         this.user = user;
         this.category = category;
@@ -37,11 +37,11 @@ public class PostModel {
         this.location = location;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
